refactor(graphql): move sentryConfig → utils, import sentry-init.mjs 1st
refs US-171 (trying to make graphql/src/index.ts more manageable)
